+++ src/components/Common/PostsWrapper/index.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-export default styled.section`
+const PostsWrapper = styled.section`
   margin: auto;
   padding: 120px 0 0;
   max-width: 1176px;
@@ -26,3 +26,5 @@ export default styled.section`
     font-size: 14px;
   }
 `;
+
+export default PostsWrapper;