+++ src/components/Footer/index.jsx
@@ -1,18 +1,17 @@
 import React from 'react';
 import styled from 'styled-components';
-import { PRIMARY_COLOR } from '~/components/Common/constants';
 
 const FooterWrapper = styled.section`
   padding: 10px 0;
-  color: #fff;
-  background-color: ${PRIMARY_COLOR};
+  color: #000;
+  background-color: #eee;
   font-size: 11px;
   font-weight: 500;
   text-align: center;
 
   a {
-    color: #fff;
-    text-decoration: underline;
+    color: #000;
+    font-weight: 700;
   }
 `;