+++ src/pages/404.jsx
@@ -1,9 +1,11 @@
 import React from 'react';
 import Link from 'gatsby-link';
+import Helmet from 'react-helmet';
 import SimpleWrapper from '~/components/Common/SimpleWrapper';
 
 const Wrapper = SimpleWrapper.extend`
   padding: 100px 0 0;
+  min-height: 100vh;
   text-align: center;
   @media (max-width: 414px) {
     padding: 70px 0 0;
@@ -21,6 +23,10 @@ const Wrapper = SimpleWrapper.extend`
 
 const NotFoundPage = () => (
   <Wrapper>
+    <Helmet>
+      <title>WONISM | PAGE NOT FOUND</title>
+      <meta name="og:title" content="WONISM | PAGE NOT FOUND" />
+    </Helmet>
     <h1>Page Not Found</h1>
     <Link to="/">← Go Home</Link>
   </Wrapper>