+++ src/store/app/sagas.js
@@ -18,6 +18,7 @@ export function* historyGoBack() {
 
     yield put({ type: HISTORY_GO_BACK_SUCCESS });
   } catch (e) {
+    yield call(() => { alert('can not access the history API'); }); // eslint-disable-line no-alert
     yield put({ type: HISTORY_GO_BACK_FAILED });
   }
 }