+++ src/store/posts/initializer.js
@@ -1,10 +1,6 @@
-import { INIT_COPY, LOAD_DISQUS_SCRIPT } from './actionTypes';
+import { INIT_COPY } from './actionTypes';
 
 export default (store) => {
-  store.dispatch({
-    type: LOAD_DISQUS_SCRIPT,
-  });
-
   store.dispatch({
     type: INIT_COPY,
   });