+++ src/templates/Post.jsx
@@ -7,10 +7,7 @@ import Helmet from 'react-helmet';
 import FaTags from 'react-icons/lib/fa/tags';
 import fp from 'lodash/fp';
 import {
-  historyGoBack,
-} from '~/store/app/actions';
-import {
-  initDisqusConfig,
+  loadDisqus,
   renderTweets,
   renderComponents,
   createCopyButton,
@@ -37,6 +34,7 @@ const Tags = styled.div`
 
 const PostContent = styled.section`
   padding: 1em 0 4em;
+  line-height: 1.6em;
 
   h2 {
     margin: 24px 0 10px;
@@ -75,8 +73,9 @@ const PostContent = styled.section`
 
 const ImageWrapper = styled.figure`
   position: relative;
-  width: 100%;
+  margin: 0 0 48px;
   padding: 56.25% 0 0;
+  width: 100%;
   overflow: hidden;
 
   img {
@@ -91,11 +90,12 @@ const ImageWrapper = styled.figure`
   }
 `;
 
+/* eslint-disable global-require, import/no-dynamic-require */
 class Post extends PureComponent {
   static propTypes = {
     data: PropTypes.shape({ date: PropTypes.object }).isRequired,
     location: PropTypes.shape({}).isRequired,
-    initDisqusConfig: PropTypes.func.isRequired,
+    loadDisqus: PropTypes.func.isRequired,
     renderTweets: PropTypes.func.isRequired,
     renderComponents: PropTypes.func.isRequired,
     createCopyButton: PropTypes.func.isRequired,
@@ -107,8 +107,9 @@ class Post extends PureComponent {
     const { pathname: identifier } = location;
     const url = fp.add(SITE_URL, identifier);
     const title = fp.get('data.markdownRemark.frontmatter.title')(this.props);
+    console.log('@@@@@@@@@');
 
-    this.props.initDisqusConfig({
+    this.props.loadDisqus({
       url,
       identifier,
       title,
@@ -127,28 +128,31 @@ class Post extends PureComponent {
 
   render() {
     const { data } = this.props;
-    const post = fp.get('markdownRemark')(data);
-    const siteTitle = fp.get('site.siteMetadata.title')(data);
-    const title = `${fp.get('frontmatter.title')(post)} | ${siteTitle}`;
-    const tags = fp.get('frontmatter.tags')(post);
-    const image = fp.flow(fp.get('frontmatter.images'), fp.first)(post);
+    const post = fp.get('markdownRemark.frontmatter')(data);
+    const { title, tags, date, images } = post;
+    const image = fp.first(images);
 
     return (
       <PostWrapper>
         <Helmet>
-          <title>{title}</title>
-          <meta name="og:title" content={title} />
+          <title>
+            WONISM | {title}
+          </title>
+          <meta name="og:title" content={`WONISM | ${title}`} />
         </Helmet>
         {fp.isNil(image) ? null : (
           <ImageWrapper>
-            <img src={image} alt={title} />
+            <img
+              src={fp.includes('//')(image) ? image : require(`~/resources/${image}`)}
+              alt={title}
+            />
           </ImageWrapper>
         )}
         <h1>
-          {fp.get('frontmatter.title')(post)}
+          {title}
         </h1>
         <time>
-          {fp.flow(fp.get('frontmatter.date'), formattedDate)(post)}
+          {formattedDate(date)}
         </time>
         {fp.isEmpty(tags) ? null : (
           <Tags>
@@ -177,12 +181,12 @@ class Post extends PureComponent {
     );
   }
 }
+/* eslint-enable global-require, import/no-dynamic-require */
 
 export default connect(
   state => state,
   {
-    historyGoBack,
-    initDisqusConfig,
+    loadDisqus,
     renderTweets,
     renderComponents,
     createCopyButton,
@@ -192,12 +196,6 @@ export default connect(
 /* eslint-disable no-undef */
 export const pageQuery = graphql`
   query PostByPath($path: String!) {
-    site {
-      siteMetadata {
-        title
-        author
-      }
-    }
     markdownRemark (
       frontmatter: { path: { eq: $path } }
     ) {